requirements.txt
bitstring==3.1.6
paho_mqtt==1.4.0
pyserial==3.4
